requirements.txt
Flask
Flask-Login
Flask-SQLAlchemy
psycopg2-binary
cx_oracle
pandas
python-dotenv
schedule
Werkzeug
